+++ docs_src/reffk/example1.py
@@ -1,0 +1,22 @@
+import edgy
+from edgy import Database, Registry
+
+database = Database("sqlite:///db.sqlite")
+models = Registry(database=database)
+
+
+class User(edgy.Model):
+    id: int = edgy.IntegerField(primary_key=True)
+    name: str = edgy.CharField(max_length=255)
+
+    class Meta:
+        registry = models
+
+
+class Post(edgy.Model):
+    id: int = edgy.IntegerField(primary_key=True)
+    user: User = edgy.ForeignKey(User)
+    comment: str = edgy.TextField()
+
+    class Meta:
+        registry = models