+++ docs_src/reffk/references.py
@@ -1,0 +1,6 @@
+from edgy import ModelRef
+
+
+class PostRef(ModelRef):
+    __model__ = "Post"
+    comment: str